+++ CHANGES.md
@@ -114,6 +114,8 @@ MmWaveEnbMac
 * The number of DL and UL CTRL symbols can be configured now in MmWaveMacSchedulerNs3
 through the attributes DlCtrlSymbols and UlCtrlSymbols.
 
+* GetN2Delay moved to MmWaveEnbPhy.
+
 ### Changed behavior:
 
 * K0, K1, K2 Delays are removed from the phy-mac common, instead they are